Avago 93xx controllers have differing firmware/driver versions.
Ticket:	#15531

For the 92xx controllers the firmware version on the controller
must match the driver version the OS uses.

If you thought that would be the same for the 93xx controllers
you would be wrong.  In the 93xx controller line the driver
version is one major version number ahead of the firmware
version number.

So a version 13 driver needs the P12 firmware.
This extends all the way back to the orignal driver for
this card, which was called version 02.00.00.00, just so the
firmware used could be version 01.00.02.00.

Younger me would likely put something in this commit message
about what a bad idea that is, using politically incorrect
parlance, and calling in to question the intellence, morals,
and possibly ethnic heritage of the people that made this
decision.

Current me is beyond all that but makes the time to type out
this commit message.

Future me would likely stop at line 5.  At least I hope that's
what he'd do.

(cherry picked from commit f059da29f48218f2fbbbb102071d3d03b5eeff0b)